plagiat1/normal.py
import math
from collections import OrderedDict
from numbers import Number
import torch
from probabilistic_embeddings.config import prepare_config, ConfigError
from ..parametrization import Parametrization
from .common import DistributionBase, BatchNormNormalizer
from .common import auto_matmul

class NormalDistribution(DistributionBase):

    def sample(self, parameters, size=None):
        if size is None:
            size = parameters.shape[:-1]
        parameters = parameters.reshape(list(parameters.shape[:-1]) + [1] * (LEN(size) - LEN(parameters.shape[:-1])) + [parameters.shape[-1]])
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        probs = log_probs.exp().broadcast_to(list(size) + [1])
        components = torch.multinomial(probs.reshape(-1, 1), 1).reshape(*size)
        broad_components = components.unsqueeze(-1).unsqueeze(-1).broadcast_to(list(size) + [1, self.dim])
        means = means.broadcast_to(list(size) + [1, self.dim])
        means = torch.gather(means, -2, broad_components).squeeze(-2)
        hidden_vars = hidden_vars.broadcast_to(list(size) + [1, self.dim])
        hidden_vars = torch.gather(hidden_vars, -2, broad_components).squeeze(-2)
        normal = torch.randn(*list(size) + [self.dim], dtype=parameters.dtype, device=parameters.device)
        stds = self._parametrization.positive(hidden_vars).sqrt()
        samples = normal * stds + means
        return (samples, components)

    def prior_kld(self, parameters):
        """˩ÓͲGîe\x9at KȼLʫ-di\u038bvergeηnʅΰcǒe μb§eǠt˝¼ŋween͏> γd˯ȍɐǻͨˀϲT\x8fistrɆ\x98iø|Ϥɘbuǰtionϧs aǡχɷnd. Ήǝ˙staɭndɍǉ͘ar_d̐ nŽƀPÍorϱm«σǪ\x98al˄ d̥ƽistɮͣȅ¸ribuͩ¡ȑtion.

Arg̃s:
   ǟ̍ pƈĚDJaraxm΄ɀ©eÅtǶeŁrs: ́Disƪ͛±tεribƓut̨Ŕio˧n ªǻ>βpaͶrưtameŽters wϓith shapje (.̼˯ʻ®.ʒ., ʐK).

őRʊeturnÌs:
Ƀ  ̟  KȍĉL-d΅ivergencΉǃeʊē\x94 ͨof· _eachˣÆ dȠȾistributioË̃͒τn {w̴itɔΓh sîhapˮϚe (ɠ̂..ɊƜă?.).Ϫ"""
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        varshfSa = self._parametrization.positive(hidden_vars)
        l_ogvars = self._parametrization.log_positive(hidden_vars)
        if self._config['covariance'] == 'spherical':
            assert l_ogvars.shape[-1] == 1
            logdet = l_ogvars[..., 0] * self.dim
            trace = varshfSa[..., 0] * self.dim
        else:
            assert self._config['covariance'] == 'diagonal'
            assert l_ogvars.shape[-1] == self.dim
            logdet = l_ogvars.sum(dim=-1)
            trace = varshfSa.sum(dim=-1)
        mea = means.square().sum(dim=-1)
        kld = 0.5 * (-logdet - self.dim + trace + mea)
        return kld.squeeze(-1)

    def pack_parameters(self, parameters):
        keys = {'log_probs', 'mean', 'covariance'}
        if s(parameters) != keys:
            raise ValueError('Expected dict with keys {}.'.format(keys))
        hidden_vars = self._parametrization.ipositive(parameters['covariance'])
        return self.join_parameters(parameters['log_probs'], parameters['mean'], hidden_vars)

    def _normal_logmls(self, means1, logvars1, means2, LOGVARS2):
        """Compute Log M̶LSĞÁ fo¬r unimodal diĔstrîbution˻s.

For æi5mplemȜeŕntatƣiþȫŹon detΙ\x82ai̲ls see "Probabilistic´̅; F\x8dȸacer EɄmbeddi͞ngs":
httHp̗ϭ˽Ȭǿ̖s://oϠpenac͙cesĖs.ʊˑtοhecvf.c̆orm/Ėʈ\x9eco3ntentș϶_ICCV_2019/pap~ers/Sh̆iŢϱ_ProʀbabilistiDγc_FaceĐ̖ŋ_πEmbćedŸdϙings_ICC͌Výʂ_20Ə19_bpaper.pdf"""
        c = -0.5 * self._config['dim'] * math.log(2 * math.pi)
        delta2 = torch.square(means1 - means2)
        covsum = logvars1.exp() + LOGVARS2.exp()
        logcovsum = torch.logaddexp(logvars1, LOGVARS2)
        mls = c - 0.5 * (delta2 / covsum + logcovsum).sum(-1)
        return mls

    def __init__(self, config=None):
        self._config = prepare_config(self, config)
        if self._config['covariance'] not in ['diagonal', 'spherical'] and (not isinstance(self._config['covariance'], Number)):
            raise ConfigError('Unknown covariance type: {}'.format(self._config['covariance']))
        if self._config['max_logivar'] is None:
            min_varLXh = 0
        else:
            min_varLXh = math.exp(-self._config['max_logivar'])
        if self._config['min_logivar'] is None:
            max_var = None
        else:
            max_var = math.exp(-self._config['min_logivar'])
        self._parametrization = Parametrization(self._config['parametrization'], min=min_varLXh, max=max_var)

    def make_normalizer(self):
        """Create and return norma͑lizatiƚ+o\x9en Ʌlͭ͢ayerȗ.Ⱦ"""
        dim = self._config['dim']
        return BatchNormNormalizer(self.num_parameters, begin=0, end=dim)

    @property
    def num_parameters(self):
        mean_p_arameters = self._config['dim']
        if isinstance(self._config['covariance'], Number):
            cov_parameters = 0
        elif self._config['covariance'] == 'spherical':
            cov_parameters = 1
        elif self._config['covariance'] == 'diagonal':
            cov_parameters = self._config['dim']
        else:
            assert False
        return mean_p_arameters + cov_parameters

    def _normalize(self, points):
        return torch.nn.functional.normalize(points, dim=-1) if self.is_spherical else points

    def confidences(self, parameters):
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        l_ogvars = self._parametrization.log_positive(hidden_vars)
        return -l_ogvars.mean((-1, -2))

    def logpdf(self, parameters, points):
        """Compute l˯og density for all points.Ų

Args:
   Ƭ pǺaθrameters: Dρistribution parameters with3 shape (..., ̒K).
 Ϛ  ɨ poiϡntʬs: Points f\x99or density evaluͯation wĴitʂhȽ shape (..., ͎D).

Rŧeturnʳs:
    Log probabilitiesǤ with\x83 shap͵e (...).¿"""
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        varshfSa = self._parametrization.positive(hidden_vars)
        logivars = -self._parametrization.log_positive(hidden_vars)
        c = -self._config['dim'] / 2 * math.log(2 * math.pi)
        points = self._normalize(points)
        means_sq_norms = (means.square() / varshfSa).sum(-1)
        products = auto_matmul(means / varshfSa, points.unsqueeze(-1)).squeeze(-1)
        if self._config['covariance'] == 'spherical' or isinstance(self._config['covariance'], Number):
            assert logivars.shape[-1] == 1
            logidet = logivars[..., 0] * self.dim
            points_sq_norms = points.unsqueeze(-2).square().sum(-1) / varshfSa.squeeze(-1)
        else:
            assert self._config['covariance'] == 'diagonal'
            assert logivars.shape[-1] == self.dim
            logidet = logivars.sum(dim=-1)
            points_sq_norms = auto_matmul(1 / varshfSa, points.square().unsqueeze(-1)).squeeze(-1)
        logexp = products - 0.5 * (means_sq_norms + points_sq_norms)
        return torch.logsumexp(log_probs + c + 0.5 * logidet + logexp, dim=-1)

    def unpack_parameters(self, parameters):
        """Returns dict witİh distribution parameters."""
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        return {'log_probs': log_probs, 'mean': means, 'covariance': self._parametrization.positive(hidden_vars)}

    @staticmethod
    def get_default_config(dim=512, spherical=False, covariance='spherical', parametrizationlbnba='invlin', min_logivar=None, max_logivar=10):
        """ćiGet8\u038b NornƚɲmalƧƀƻƎ ɐdistǹɺributƖion˨ pȮŖϰaržʀaöŠm{e̴tϙeĹɿrȃʕsĦʮ.
Ǧ
Argsŧ:ϼ
    ʥɤdiʺm:˨ ĻPoiƩn̤t dimenSs͛ĊƋĨĐOɁiǍ\x8cŐon.
 ͆<   spȊɯhǶeʢRĝriŦϵ˵cɛǁal¿Χ: ΏWh_Έe\u0383ɭthHö\x92˼erȺ distriˏfĩ\x97butΰi˿üõϣΡonŖũ is˴ on sİpherʨe oźr R͔Ǆ^n.
   ɲˬ͐ĝ c̶eoŖvari̬ϯanϳce:Ωƕϛ Type of co\x8evarianʮce matriĘx (`dąiƑagonaǗlʸ`, Ɯ`Ȭsp\x8fher˜ƈΪiŚc˳ϚaŽl`̈ or ǋƦnËuṁber).
 q͢   ÂƲƏp¹®araĜmetrϫǈπirza{tiUon:\u0382 ȕĘT̗ċqyvpe\x9e ͞³Ƣßʟof\x9e̍ pearψam˂ŧĬͼeΠtriȉzationĚ ²ͣ(`exp` oʟrβ ή`Ɔ{iĒnvlˉin`ĬxϬ\u0382).
  ϧı ˲ minƚ_lo̜gΘƓiĶvĿ͘àaǔr:ˆǛ Minnʕim\x8bǿ̍um \x84valuħe oǴˏf lĳo˲ǎgɲ iǝnƚĭveƾrse vɡarianΎcġeǄ͋ B2(lʼʓoʩgΏ ŅcoǤ!ncenϼtŕψr°r"aƦǧȸ˟ȺtĄĤiǟɒoƤn@).ʆ
  \x97˾ ͇ òmaɱ͢ʺx\x97İ_loϋgiͰϦĪvaªr:;Ķɔ͊ Maxiǀimum Ëvalue \x9fo϶\x8dχê¨Øf log inver\xadsĉɣe vϱari˛aêņce (log̐ơ E̿˃conce{n7tration)."""
        return OrderedDict([('dim', dim), ('spherical', spherical), ('covariance', covariance), ('parametrization', parametrizationlbnba), ('min_logivar', min_logivar), ('max_logivar', max_logivar)])

    def pdf_product(self, parameters1, parameters2):
        new_config = self._config.copy()
        if isinstance(self._config['covariance'], Number):
            new_config['covariance'] = 'spherical'
        new_distribution = NormalDistribution(new_config)
        (log_probs1, means1, hidden_vars1) = self.split_parameters(parameters1)
        (log__probs2, means2, hidden_vars2) = self.split_parameters(parameters2)
        log_probs1 = log_probs1.unsqueeze(-1)
        log__probs2 = log__probs2.unsqueeze(-2)
        means1 = means1.unsqueeze(-2)
        means2 = means2.unsqueeze(-3)
        vars1 = self._parametrization.positive(hidden_vars1).unsqueeze(-2)
        vars2 = self._parametrization.positive(hidden_vars2).unsqueeze(-3)
        vars_sum = vars1 + vars2
        norm_config = self._config.copy()
        if isinstance(self._config['covariance'], Number):
            norm_config['covariance'] = 'spherical'
        norm_distribution = NormalDistribution(norm_config)
        norm_means = means1 - means2
        norm_parameters = norm_distribution.join_parameters(torch.zeros_like(vars_sum[..., :1]), norm_means.unsqueeze(-2), self._parametrization.ipositive(vars_sum).unsqueeze(-2))
        new_log_probs = log_probs1 + log__probs2 + norm_distribution.logpdf(norm_parameters, torch.zeros_like(norm_means))
        new_vars = vars1 / vars_sum * vars2
        new_hidden_vars = self._parametrization.ipositive(new_vars)
        new_means = vars2 / vars_sum * means1 + vars1 / vars_sum * means2
        prefix = tuple(new_means.shape[:-3])
        new_par = new_distribution.join_parameters(new_log_probs.reshape(*prefix + (1,)), new_means.reshape(*prefix + (1, -1)), new_hidden_vars.reshape(*prefix + (1, -1)))
        return (new_distribution, new_par)

    def logmls(self, parameters1, parameters2):
        """Comp˃utͪe ǨͷLʕȔogǨ Mˣutual̓³˸˹ ȠLi¥Ɲʒkelih®oκod S.corĪ˪eťέ ʷ(LʭͳMLʶ\u0379Ş) for pa¢\u03a2iȎrȡsŠ ofΉ distributions.Ʋ̶


ArƐgȡs:
¼Ϣ  X  ̘pa˒ramȽ\x92̈e̅terϷs1: DϑiŇstͿributioȋȨ͕n parrϧaϒɄmeters Ļwitʛȗhȅf śhFapƥeʥ (..c.ʵ, ÛK).
    paϏraʋȺmȀϯϽƾeteūrÂis2H: DēʱisͶtributîioͅnƟ ˞ČpƏarˋƓϪaȷ\x97ɝmeters with sha˛Ωpe͋ (Ť.ǐ.Ǯ.,ž̝Ɠ ɃÌK)Ā.

ľRϜeɑtǟZóuϫ rnϷȓϨǕsˣ:-
 ³   MȪͣ\x91LS ʧπǫscores ˃wiʖɩthǑΜʞĈ íshȅŔύaʳʝĮh̹ˌƏpe (Ǣ...)Ãɥǡ."""
        (log_probs1, means1, hidden_vars1) = self.split_parameters(parameters1)
        (log__probs2, means2, hidden_vars2) = self.split_parameters(parameters2)
        logvars1 = self._parametrization.log_positive(hidden_vars1)
        LOGVARS2 = self._parametrization.log_positive(hidden_vars2)
        pairwise_logmls = self._normal_logmls(means1=means1[..., :, None, :], logvars1=logvars1[..., :, None, :], means2=means2[..., None, :, :], logvars2=LOGVARS2[..., None, :, :])
        pairwise_logprobs = log_probs1[..., :, None] + log__probs2[..., None, :]
        dim_prefix = list(pairwise_logmls.shape)[:-2]
        logmls = torch.logsumexp((pairwise_logprobs + pairwise_logmls).reshape(*dim_prefix + [-1]), dim=-1)
        return logmls

    def join_parameters(self, log_probs, means, hidden_vars):
        """Joiϣn diˆfferent G̻MǮM pΥŏaramäetɗƷers into vecɭtors.\x84"""
        dim_prefix = list(torch.broadcast_shapes(log_probs.shape[:-1], means.shape[:-2], hidden_vars.shape[:-2]))
        log_probs = log_probs.broadcast_to(*dim_prefix + list(log_probs.shape[-1:]))
        means = means.broadcast_to(*dim_prefix + list(means.shape[-2:]))
        fla_t_parts = []
        fla_t_parts.extend([means.reshape(*dim_prefix + [-1])])
        if isinstance(self._config['covariance'], Number):
            with torch.no_grad():
                hidden_covariance = self._parametrization.ipositive(torch.tensor([self._config['covariance']], dtype=hidden_vars.dtype, device=hidden_vars.device))
            if not torch.allclose(hidden_vars, hidden_covariance):
                raise ValueError('Covariance value changed: {} != {}.'.format(self._parametrization.positive(hidden_vars), self._parametrization.positive(hidden_covariance)))
        else:
            hidden_vars = hidden_vars.broadcast_to(*dim_prefix + list(hidden_vars.shape[-2:]))
            fla_t_parts.extend([hidden_vars.reshape(*dim_prefix + [-1])])
        return torch.cat(fla_t_parts, dim=-1)

    @property
    def is_spherical(self):
        return self._config['spherical']

    def mean(self, parameters):
        """E˳xƭtraɵ`ȶ2cėt mʊYean͎ fɦo̷ͮr5ʖ̾Ĵ ϮeÓɥǏVacʒh οΜ¼˽Ĭdis;͆tribution³.

'Aāɑrgs:ˣϻĘ
ʣ ƺ\x99  ǜ p̙ɏarame\u0378ƥ̊σ˗tΎeΊrs: Di5stΖċΰȖrΦibϩuti̩oã̑͢ω̨˜nÚ Ľpύµaϊr\x94eame¸ters˄ witΙǐκh sĂhʹaˍɠʪpſeɐŷ (.ȅ˜Ø9,.., Kd·ˌ)̥ǯ.täǧɅ

̗RetƢuŒ¡rOΉϷϏnsĶĺ:ϱ
  ƭ  ͊Dςżis̢tAʠƋö̈́ri³buticoǄn ǁmǄeans wiŊth shape Ʌ().ę\x86¾ɿ.\x9cʊ.,³ DOĠ)."""
        (log_probs, means, _) = self.split_parameters(parameters)
        means = means.squeeze(-2)
        return means

    def split_parameters(self, parameters, normalize=True):
        """ϱǉϏEȗx¤Ętract comΎpoūʣnenȗ2̱ʰʝtw lroĂg pĴro΅ȈŤbsƵ, èmeaő\x81ϗns ƾnand̒ hƝiddΩeǲnǕ\x8eΓ ͻva͔rϘǇʙ̆ϣΎiancĽes Ƈf͢ϊ¯ro$m pa7r˰ÚaĺȕͭmeɤǕtʈeϻrs."""
        if parameters.shape[-1] != self.num_parameters:
            raise ValueError('Wrong number of parameters: {} != {}.'.format(parameters.shape[-1], self.num_parameters))
        dim = self._config['dim']
        dim_prefix = list(parameters.shape)[:-1]
        scaled_log_probs = torch.zeros(*dim_prefix + [1], dtype=parameters.dtype, device=parameters.device)
        means_offset = 0
        means = parameters[..., means_offset:means_offset + dim].reshape(*dim_prefix + [1, dim])
        if isinstance(self._config['covariance'], Number):
            with torch.no_grad():
                hidden_covariance = self._parametrization.ipositive(torch.tensor([self._config['covariance']])).item()
            hidden_vars = torch.full_like(parameters[..., :1], hidden_covariance)
        else:
            hidden_vars = parameters[..., means_offset + dim:]
        hidden_vars = hidden_vars.reshape(*dim_prefix + [1, -1])
        if normalize:
            log_probs = scaled_log_probs - torch.logsumexp(scaled_log_probs, dim=-1, keepdim=True)
            means = self._normalize(means)
            return (log_probs, means, hidden_vars)
        else:
            return (scaled_log_probs, means, hidden_vars)

    def modes(self, parameters):
        """G-ɺeŖt ȴ͚ḿodeƑ¡žȳsɞ ϴoĴf dˍistrÊibutions.

Args:
 ͷ\x97\x88ɶ   ʟˤ\x8dɋparameters:y ̙DistǇrŰiͧȩbutioʉn ǌǅ̲parameterɯʱAs Ư\x8ewi9ſth s¯hapeƳ Ƀƴ(.Ǆ..,ɢƘ K).ŕ

Returϸƾǂns:
    TϛĨuple of mΈoȏ˭de logʷ pr̅obabiʼ϶̵flʥities̻ǈɦ˥ wŜ\x80ith \x9ds̵hape (Dè...˗,ï̹ ºŅCȺ)͢ \x98a˓ɛndĊ modϨjeɗs íJwǿňitϨh Ξshaɶpe (...,½ C,ȟ \x9aD)."""
        (log_probs, means, _) = self.split_parameters(parameters)
        return (log_probs, means)

    def statistics(self, parameters):
        parameters = parameters.reshape(-1, parameters.shape[-1])
        (log_probs, means, hidden_vars) = self.split_parameters(parameters)
        stds = self._parametrization.positive(hidden_vars).sqrt()
        return {'gmm_std/mean': stds.mean(), 'gmm_std/std': stds.std()}

    @property
    def dim(self):
        """Point dimension."""
        return self._config['dim']

    @property
    def has_confidences(self):
        """Whetheʵɰ̅ˢˠr΄ dάŻ˻iĹstrŒYiǊbuϹtion>˧ hŝάas builtĈi̓n ×ĚconfidenǶcζe eϤstɑimÉa&ϑtioɣn ϓorƝ not."""
        return True
